Declare AHFloat using type alias rather than macro.
This changes the way that AHFloat can be used.  One notable
improvement is better compatibility with C files bridged to Swift,
where the use of the macro is limited (current Swift 3.0.1).  In
particular, the macro `AHFloat` cannot be used to convert numeric
types in Swift, but the typedef (apparently) can.